requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.8.3
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
beautifulsoup4==4.11.1
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
dj-database-url==1.0.0
    # via django-heroku
django==4.1.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-filter
    #   django-heroku
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-heroku==0.3.1
    # via -r requirements.in
django-stubs-ext==0.7.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-analytics-data==0.14.2
    # via -r requirements.in
google-api-core[grpc]==2.10.2
    # via google-analytics-data
google-auth==2.14.1
    # via google-api-core
googleapis-common-protos==1.57.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.50.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.50.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.22.1
    # via google-analytics-data
protobuf==4.21.9
    # via
    #   google-analytics-data
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.5
    # via django-heroku
psycopg2-binary==2.9.5
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via google-api-core
rsa==4.9
    # via google-auth
sentry-sdk==1.11.1
    # via -r requirements.in
six==1.16.0
    # via
    #   google-auth
    #   grpcio
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
typing-extensions==4.4.0
    # via django-stubs-ext
uritemplate==4.1.1
    # via -r requirements.in
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
whitenoise==6.2.0
    # via django-heroku
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
